Moved files to correct packages & deleted unused code.
