+++ app/src/main/java/se/ju/student/robomow/data/CollisionAvoidanceImage.kt
@@ -1,4 +1,4 @@
-package se.ju.student.robomow.model
+package se.ju.student.robomow.data
 
 data class CollisionAvoidanceImage(
     val imageURL: String
