--- app/src/main/java/se/ju/student/robomow/ui/TestParcelableActivity.kt
@@ -1,27 +1,0 @@
-package se.ju.student.robomow.ui
-
-import android.os.Build
-import androidx.appcompat.app.AppCompatActivity
-import android.os.Bundle
-import se.ju.student.robomow.R
-import se.ju.student.robomow.model.MowSession
-
-class TestParcelableActivity : AppCompatActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_test_parcelable)
-        val mowSession = getMowSession()
-        for (collision in mowSession!!.avoidedCollisions) {
-            println(collision.imageLink)
-        }
-    }
-
-    private fun getMowSession(): MowSession? {
-        val mowSession = if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.TIRAMISU) {
-            intent.getParcelableExtra("MOW_SESSION", MowSession::class.java)
-        } else {
-            intent.getParcelableExtra("MOW_SESSION")
-        }
-        return mowSession
-    }
-}
